server/requirements.txt
alembic==1.13.2
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
black==24.4.2
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
coverage==7.5.3
crcmod==1.7
decorator==5.1.1
Django==5.0.1
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
executing==2.0.1
factory_boy==3.3.3
Faker==36.1.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.0.3
gspread==6.1.2
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
pygraphviz==1.14
PyJWT==1.7.1
pyparsing==3.1.2
pytest==8.2.2
pytest-django==4.10.0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.18.1
rsa==4.9
shortuuid==1.0.13
six==1.16.0
SQLAlchemy==2.0.31
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
